wx_tests v0.1.0 ElixirToHtml

Functions

colour(token, colour)

Vgiven a token and a colour, return the necessary HTML

colourize(tokens)

Iterate over the list of tokens returned by lex, turning them into HTML

colourize(list, html, flags)

encode_entities(charlist)

Encode the HTML entities so that the dont screw up the HTML representation.

fileToHtml(inputFile)

fileToHtmlFile(inputFile, outputFile \\ :nofile)

Colourise the contents of the input file, checking that the input file is OK.

getFiles(input)

Given the input file from the command line, normalize it, and generate an html file name for the output.

lex(s)

Interface to leex

parse(inputFile, outputFile)

Colourize the given file, returning an HTML string

stringToHtml(code)