wx_tests v0.1.0 WxDsl

An implementation of a DSL for generating GUI’s

Functions

bgColour(colour) (macro)

Set the background colour for the enclosing control.

The colour may be either one of the defined colours in wxDefines.ex or a nummeric RGB specification in the form {r, g, b}.

  bgColour(@wxSILVER)
  bgColour({{192, 192, 192}})

border(attributes) (macro)

boxSizer(attributes, list) (macro)

Create a new box sizer.

button(attributes) (macro)

button(attributes, list) (macro)

codeWindow(attributes) (macro)

codeWindow(attributes, list) (macro)

event(eventType) (macro)

event(eventType, callBack) (macro)

events(attributes) (macro)

Macro to set up the event connections for the window. The attributes consist of one of the following: {: callback | nil, : value, …}

event may be one of: :command_button_clicked :close_window :timeout # This is a function to be called repeatedly every

                        # n seconds where n is given by the :delay options
                        # in milliseconds (default 1000 (1 second))

if a callback is supplied it must have arity 4.

frame(attributes, list) (macro)

get_opts_map(opts)

get_opts_map(list, opts)

htmlWindow(attributes) (macro)

htmlWindow(attributes, list) (macro)

layout(attributes) (macro)

layout(layout, attributes) (macro)

mainWindow(attributes, list) (macro)

This is the top level window and should be the outer element of a window specification. This window performs the initialisation.

Parameter	Description	Value	Default

name	The name by which the window will be referred to.	atom()	modulename

show	If set to true the window will be made visible when the construction is complete. If set to false, the window will be invisible until explicitly shown using :wxFrame.show(frame).	Boolean()	true

Example:

defmodule TestWindow do
  use WxDsl
  import WxDefines

  def createWindow(show) do
    mainWindow show: show do
      # Create a frame with a status bar and a menu.
      frame id: :main_frame,
      ...
      ...
      end
    end
  end

menu(attributes, list) (macro)

menuBar(list) (macro)

menuItem(attributes) (macro)

menuSeparator() (macro)

panel(attributes, list) (macro)

replaceIfTrue(originalValue, condition, newValue)

scrolledWindow(attributes, list) (macro)

setEvents(events)

setEvents(list, parent)

spacer(attributes) (macro)

staticBoxSizer(attributes, list) (macro)

Create a new box sizer.

staticText(attributes) (macro)

statusBar(attributes) (macro)

Add a status bar to the enclosing frame.

attributes	Description	Value	Default

number	The number of fields to create. Specify a value greater than 1 to create a multi-field status bar.	atom()	modulename

style	The status bar style.	integer()	1

text	The initial status bar text	string()

                                                                          list of strings        | []                     |

Example:

statusBar(title: "ElixirWx Menu Test")

textControl(attributes) (macro)

tool(attributes) (macro)

toolBar(attributes, list) (macro)

window(attributes) (macro)

window(attributes, list) (macro)

Create a new window.

attributes	Description	Value	Default

style	The window Style.	atom()	modulename

size	The initial size of the window.	atom()	modulename

layout	The window layout.	list	[]

Example:

window(style: @wxBORDER_SIMPLE, size: {50, 25},
      layout: [proportion: 0, flag: @wxEXPAND]) do
  bgColour(@wxBLACK)
end

xstaticBoxSizer(attributes, list) (macro)