wx_tests v0.1.0 WxStatusBar

A status bar is a narrow window that can be placed along the bottom of a frame to show small amounts of status information.

A status bar can contain one or more fields, one or more of which can be variable length according to the size of the window.

The status bar also maintains an independent stack of status texts for each field (see pushStatusText() and popStatusText()).

Functions

new(parent, attributes)

Create a new status bar and attach it to the main frame. If a :text option is supplied, set the text.

setText(text)

Set the status bar text. If the suppied test is a string, then set it in the first field. If it is a list of strings, then set all fields, setting the number of fields to the length of the supplied list

setText(text, index)