wx_tests v0.1.0 WxUtilities

General utilities

Functions

getObjOpts(options, defaults)

Given a list of otions, remove the ones not in the allowed key list. Returns {id, [options], layout, [errors]} where: id = value of the option key or :no_id options = the list of options as specified by the defaults. If a layout = content of the :layout key in the options, or []

     If layout is an atom, the content of the matching layout key is returned.

Where errors are any options not in the allowed list

getOptions(options, defaults)

Given a list of otions, remove the ones not in the allowed key list. Returns {[options], [errors]} Where errors are any options not in the allowed list