wx_tests v0.1.0 WxWinObj

This object creates, encapsulates and provides an interface to a window.

Functions

child_spec(arg)

Returns a specification to start this module under a supervisor.

See Supervisor.

getAttr(window, obj, attr)

Get the given attribute value.

window: Atom; The name of the window (see start_link()) obj: Atom; The name of the window element. attr: Atom; The name of the attribute to be retrieved

returns the value of the attribute.

getAttr(:my_window, :ok_button, :text)

hide(window)

Hide the specified window.

hide(window)

setAttr(window, obj, attr, value)

Set the given attribute value.

window: Atom; The name of the window (see start_link()) obj: Atom; The name of the window element. attr: Atom; The name of the attribute to be retrieved value: Term; The value to set the attribute to.

setAttr(:my_window, :ok_button, :size, {100, 200})

show(window)

Show the specified window.

show(pid)

start(window_spec, evt_handler, options \\ [])

See start_link()

start_link(window_spec, evt_handler, options \\ [])

Create and optionally show a window.

windowSpec: The name of the module containing the window specification (see WxDsl).

evtHandler: The name of th emodule containing the event handling code.

options: Options when creating a window.They can be:

show: Bool, Show the window when created(default)

name: The name that the window will be registered with. if a name is not supplied or is nil, then the name of the module containing the winowSpec will be used to register the window.

  start_link(MyWindow, MyWindowEvents, show: true, name: MyWindow)

statusBarText(pid, text, index \\ 0)

Set the status bar text.