wx_tests v0.1.0 WxWinObj.API

An API to make using WxWindowObject simpler, and tomake sure any errors are reported in the callers code.

Functions

closeWindow(window)

Close a window.

the window parameter may be either the name given when the window was created, or the PID returned by the new window call.

hideWindow(window)

Hide a window.

the window parameter may be either the name given when the window was created, or the PID returned by the new window call.

newWindow(windowSpec, eventHandler, options \\ [])

Create a new window and optionally show it.

windowSpec: The name of the module containing the window specification (see WxDsl).

evtHandler: The name of th emodule containing the event handling code.

options: Options when creating a window.They can be:

show: Bool, Show the window when created(default)

name: The name that the window will be registered with. if a name is not supplied or is nil, then the name of the module containing the winowSpec will be used to register the window.

  start_link(MyWindow, MyWindowEvents, show: true, name: MyWindow)

showWindow(window)

Show a window.

the window parameter may be either the name given when the window was created, or the PID returned by the new window call.

waitForWindowClose(waitWindow, timeout \\ -1)

Wait for window to close.

All events apart from the termiantion event from the specified window will be discarded. If window is nil, then a termination event from any window will cause a return.

If a timeout is supplied, then waitForTermination will return after the specified number of seconds.

Returns:

{windowName, :window_closed, reason}

:timeout