wx_tests v0.1.0 WxWinObject

This object creates, encapsulates and provides an interface to a window.

Functions

child_spec(arg)

Returns a specification to start this module under a supervisor.

See Supervisor.

hide(window)

Hide the window.

hide(window)

show(window)

Show the window.

show(pid)

start(window_spec, evt_handler, options \\ [])

See start_link()

start_link(window_spec, evt_handler, options \\ [])

Create and optionally show a window.

windowSpec: The name of the module containing the window specification (see WxDsl).

evtHandler: The name of th emodule containing the event handling code.

options: Options when creating a window.They can be:

show: Bool, Show the window when created(default)

name: The name that the window will be registered with. if a name is not supplied or is nil, then the name of the module containing the winowSpec will be used to register the window.

  start_link(MyWindow, MyWindowEvents, show: true, name: MyWindow)

statusBarText(pid, text, index \\ 0)

Set the status bar text.